Split pipeline collection and stream pumping into focused modules
Extract pipeline completion/collection into cuprum/_pipeline_collect.py,
pipe-result triage into cuprum/_pipeline_stream_results.py, and the
backpressure pump loop into cuprum/_streams_pump.py, keeping each seam
single-purpose and under the module size ceiling. Give the pipeline invariant
guards a distinct exception type, triage pipe failures with structural pattern
matching, and split the pipeline and pump test suites to match the new
boundaries. Share the CrossHair availability probe through _crosshair_support.
