+++ cuprum/unittests/test_pipeline.py
@@ -2,10 +2,6 @@
 
 from __future__ import annotations
 
-import asyncio
-import dataclasses as dc
-import typing as typ
-
 import pytest
 
 from cuprum import (
@@ -17,108 +13,8 @@ from cuprum import (
     scoped,
     sh,
 )
-from cuprum._testing import (
-    _READ_SIZE,
-    _PipelineWaitResult,
-    _prepare_pipeline_config,
-    _pump_stream,
-    _spawn_pipeline_processes,
-    _wait_for_pipeline,
-)
-from cuprum.sh import (
-    Pipeline,
-    PipelineResult,
-    RunOutputOptions,
-)
+from cuprum.sh import Pipeline, PipelineResult, RunOutputOptions
 from tests.helpers.catalogue import python_catalogue
-
-
-class _StubPumpReader:
-    """Stub stream reader yielding queued chunks then EOF."""
-
-    def __init__(self, chunks: list[bytes]) -> None:
-        """Initialise the stub with the supplied queued chunks."""
-        self._chunks = list(chunks)
-        self.read_calls = 0
-
-    async def read(self, _: int) -> bytes:
-        """Return the next queued chunk, or empty bytes when exhausted."""
-        self.read_calls += 1
-        await asyncio.sleep(0)
-        if not self._chunks:
-            return b""
-        return self._chunks.pop(0)
-
-
-class _StubPumpWriter:
-    """Stub stream writer recording writes, drains, and closure."""
-
-    def __init__(self, *, fail_on_drain_call: int | None = None) -> None:
-        """Initialise the stub, optionally failing on a given drain call."""
-        self.data = bytearray()
-        self.drain_calls = 0
-        self.write_calls = 0
-        self.closed = False
-        self.write_eof_calls = 0
-        self.wait_closed_calls = 0
-        self._fail_on_drain_call = fail_on_drain_call
-
-    def write(self, chunk: bytes) -> None:
-        """Buffer the chunk and record the write call."""
-        self.write_calls += 1
-        self.data.extend(chunk)
-
-    async def drain(self) -> None:
-        """Record the drain, optionally raising to simulate a broken pipe."""
-        self.drain_calls += 1
-        await asyncio.sleep(0)
-        if self._fail_on_drain_call is None:
-            return
-        if self.drain_calls == self._fail_on_drain_call:
-            raise BrokenPipeError
-
-    def write_eof(self) -> None:
-        """Record that end-of-file was signalled to the writer."""
-        self.write_eof_calls += 1
-
-    def close(self) -> None:
-        """Mark the writer as closed."""
-        self.closed = True
-
-    async def wait_closed(self) -> None:
-        """Record that the caller awaited closure of the writer."""
-        self.wait_closed_calls += 1
-
-
-class _StubSpawnProcess:
-    """Stub subprocess recording terminate, kill, and wait calls."""
-
-    def __init__(self, pid: int) -> None:
-        """Initialise the stub process with the given PID."""
-        self.pid = pid
-        self.returncode: int | None = None
-        self.stdout = None
-        self.stderr = None
-        self.stdin = None
-        self.terminate_calls = 0
-        self.kill_calls = 0
-        self.wait_calls = 0
-
-    def terminate(self) -> None:
-        """Record that the process was terminated."""
-        self.terminate_calls += 1
-
-    def kill(self) -> None:
-        """Record that the process was killed."""
-        self.kill_calls += 1
-
-    async def wait(self) -> int:
-        """Record the wait and return a default terminated exit code."""
-        self.wait_calls += 1
-        await asyncio.sleep(0)
-        if self.returncode is None:
-            self.returncode = -15
-        return self.returncode
 
 
 def test_or_operator_composes_pipeline() -> None:
@@ -220,66 +116,6 @@ def _run_test_pipeline(
         return pipeline.run_sync()
 
 
-@pytest.mark.usefixtures("stream_backend")
-def test_pipeline_run_sync_accepts_run_output_options() -> None:
-    """Pipeline.run_sync accepts ``output=RunOutputOptions`` like SafeCmd."""
-    catalogue, python_program = python_catalogue()
-    python = sh.make(python_program, catalogue=catalogue)
-    echo = sh.make(ECHO)
-
-    pipeline = echo("-n", "unified") | python(
-        "-c",
-        "import sys; sys.stdout.write(sys.stdin.read())",
-    )
-
-    with scoped(ScopeConfig(allowlist=frozenset([ECHO, python_program]))):
-        result = pipeline.run_sync(output=RunOutputOptions(capture=False, echo=False))
-
-    assert result.ok is True
-    assert result.stdout is None
-
-
-@pytest.mark.usefixtures("stream_backend")
-def test_pipeline_flat_capture_echo_kwargs_are_deprecated() -> None:
-    """The flat ``capture``/``echo`` kwargs still work but warn."""
-    catalogue, python_program = python_catalogue()
-    python = sh.make(python_program, catalogue=catalogue)
-    echo = sh.make(ECHO)
-
-    pipeline = echo("-n", "legacy") | python(
-        "-c",
-        "import sys; sys.stdout.write(sys.stdin.read())",
-    )
-
-    with (
-        scoped(ScopeConfig(allowlist=frozenset([ECHO, python_program]))),
-        pytest.warns(DeprecationWarning, match="RunOutputOptions"),
-    ):
-        result = pipeline.run_sync(capture=True, echo=False)
-
-    assert result.ok is True
-    assert result.stdout == "legacy"
-
-
-def test_pipeline_rejects_output_combined_with_flat_kwargs() -> None:
-    """Supplying both ``output`` and the deprecated flags raises ValueError."""
-    catalogue, python_program = python_catalogue()
-    python = sh.make(python_program, catalogue=catalogue)
-
-    pipeline = python("-c", "print('a')") | python(
-        "-c",
-        "import sys; sys.stdout.write(sys.stdin.read())",
-    )
-
-    with scoped(ScopeConfig(allowlist=frozenset([python_program]))):
-        with pytest.raises(ValueError, match="not both"):
-            pipeline.run_sync(output=RunOutputOptions(), capture=True)
-
-        unknown_output_kwargs: dict[str, bool] = {"captuer": True}
-        with pytest.raises(TypeError, match="unexpected keyword"):
-            pipeline.run_sync(**typ.cast("dict[str, typ.Any]", unknown_output_kwargs))
-
-
 def test_pipeline_run_streams_stdout_between_stages(stream_backend: str) -> None:
     """Pipeline.run_sync streams stdout into the next stage stdin."""
     catalogue, python_program = python_catalogue()
@@ -374,98 +210,6 @@ def test_pipeline_run_sync_failure_semantics(
         assert result.final.exit_code == 0
 
 
-def test_pump_stream_drains_per_chunk() -> None:
-    """Streaming between stages awaits drain for backpressure."""
-
-    async def exercise() -> _StubPumpWriter:
-        """Drive the pump under test and capture the resulting writer."""
-        reader = asyncio.StreamReader()
-        writer = _StubPumpWriter()
-        task = asyncio.create_task(
-            _pump_stream(reader, typ.cast("asyncio.StreamWriter", writer)),
-        )
-        payload = b"x" * (_READ_SIZE * 2 + 1)
-        reader.feed_data(payload)
-        reader.feed_eof()
-        await task
-        return writer
-
-    writer = asyncio.run(exercise())
-
-    assert bytes(writer.data) == b"x" * (_READ_SIZE * 2 + 1)
-    assert writer.write_calls == writer.drain_calls
-    assert writer.drain_calls >= 2
-    assert writer.write_eof_calls == 1
-    assert writer.closed is True
-    assert writer.wait_closed_calls == 1
-
-
-def test_pump_stream_handles_downstream_close_without_hanging() -> None:
-    """_pump_stream drains stdout even if downstream closes mid-stream."""
-
-    async def exercise() -> tuple[_StubPumpReader, _StubPumpWriter]:
-        """Pump through a writer that fails mid-stream and return both ends."""
-        reader = _StubPumpReader([b"a" * _READ_SIZE, b"b" * _READ_SIZE, b"c"])
-        writer = _StubPumpWriter(fail_on_drain_call=2)
-        await _pump_stream(
-            typ.cast("asyncio.StreamReader", reader),
-            typ.cast("asyncio.StreamWriter", writer),
-        )
-        return reader, writer
-
-    reader, writer = asyncio.run(exercise())
-
-    assert reader.read_calls == 4  # 3 chunks + EOF
-    assert (
-        writer.write_calls == 2
-    )  # drain fails on second write; third chunk never written
-    assert writer.closed is True
-    assert writer.wait_closed_calls == 1
-
-
-def test_spawn_pipeline_processes_terminates_started_stages_on_failure(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Spawn failures should terminate any already-started pipeline stages."""
-    echo = sh.make(ECHO)
-    first = echo("-n", "hello")
-    second = echo("-n", "world")
-    config = _prepare_pipeline_config(
-        capture=True, echo=False, timeout=None, context=None
-    )
-
-    spawned: list[_StubSpawnProcess] = []
-    call_count = 0
-
-    async def fake_create_subprocess_exec(
-        *_: object,
-        **__: object,
-    ) -> _StubSpawnProcess:
-        """Spawn the first stage, then fail subsequent spawn attempts."""
-        nonlocal call_count
-        call_count += 1
-        await asyncio.sleep(0)
-        if call_count == 1:
-            proc = _StubSpawnProcess(pid=12345)
-            spawned.append(proc)
-            return proc
-        raise FileNotFoundError("missing")
-
-    monkeypatch.setattr(asyncio, "create_subprocess_exec", fake_create_subprocess_exec)
-
-    async def exercise() -> None:
-        """Spawn the pipeline and assert the spawn failure propagates."""
-        with pytest.raises(FileNotFoundError):
-            await _spawn_pipeline_processes((first, second), config)
-
-    asyncio.run(exercise())
-
-    assert len(spawned) == 1
-    assert spawned[0].terminate_calls == 1
-    assert spawned[0].kill_calls == 0
-    assert spawned[0].wait_calls >= 1
-
-
 def test_pipeline_requires_at_least_two_stages() -> None:
     """Pipelines reject construction with fewer than two parts."""
     echo = sh.make(ECHO)
@@ -473,246 +217,3 @@ def test_pipeline_requires_at_least_two_stages() -> None:
 
     with pytest.raises(ValueError, match="at least two stages"):
         Pipeline((only,))
-
-
-class _StubPipelineWaitProcess:
-    """Stub process modelling exit codes and readiness for wait tests."""
-
-    def __init__(
-        self,
-        *,
-        pid: int,
-        exit_code: int,
-        ready: asyncio.Event | None = None,
-    ) -> None:
-        """Initialise the stub with its PID, exit code, and ready event."""
-        self.pid = pid
-        self.returncode: int | None = None
-        self.stdout = None
-        self.stderr = None
-        self.stdin = None
-        self.terminate_calls = 0
-        self.kill_calls = 0
-        self._exit_code = exit_code
-        self._ready = ready
-
-    def terminate(self) -> None:
-        """Record termination and unblock any pending wait."""
-        self.terminate_calls += 1
-        self._exit_code = -15
-        if self._ready is not None:
-            self._ready.set()
-
-    def kill(self) -> None:
-        """Record the kill and unblock any pending wait."""
-        self.kill_calls += 1
-        self._exit_code = -9
-        if self._ready is not None:
-            self._ready.set()
-
-    async def wait(self) -> int:
-        """Await readiness if set, then return the recorded exit code."""
-        if self._ready is not None:
-            await self._ready.wait()
-        await asyncio.sleep(0)
-        if self.returncode is None:
-            self.returncode = self._exit_code
-        return self.returncode
-
-
-async def _exercise_wait_for_pipeline(
-    exit_codes: tuple[int, int, int],
-    ready_stages: frozenset[int],
-) -> tuple[
-    _StubPipelineWaitProcess,
-    _StubPipelineWaitProcess,
-    _StubPipelineWaitProcess,
-    _PipelineWaitResult,
-]:
-    """Execute _wait_for_pipeline with stub processes and custom exit scenarios.
-
-    Parameters
-    ----------
-    exit_codes:
-        Exit code for each of the three stages.
-    ready_stages:
-        Set of stage indices that should be immediately ready.
-
-    Returns
-    -------
-    tuple
-        Tuple of (process0, process1, process2, wait_result).
-
-    """
-    events = [asyncio.Event() for _ in range(3)]
-    for idx in ready_stages:
-        events[idx].set()
-
-    processes = [
-        _StubPipelineWaitProcess(pid=i + 1, exit_code=exit_codes[i], ready=events[i])
-        for i in range(3)
-    ]
-
-    result = await _wait_for_pipeline(
-        typ.cast("list[asyncio.subprocess.Process]", processes),
-        pipe_tasks=[],
-        cancel_grace=0.01,
-        started_at=[0.0, 0.0, 0.0],
-    )
-
-    return processes[0], processes[1], processes[2], result
-
-
-def _assert_stage_terminated(
-    process: _StubPipelineWaitProcess,
-    *,
-    should_terminate: bool,
-) -> None:
-    """Assert whether a process was terminated during fail-fast."""
-    if should_terminate:
-        assert process.terminate_calls == 1, (
-            f"Process {process.pid} should be terminated"
-        )
-    else:
-        assert process.terminate_calls == 0, (
-            f"Process {process.pid} should not be terminated"
-        )
-
-
-def _assert_pipeline_failure(
-    result: _PipelineWaitResult,
-    *,
-    failure_index: int | None,
-    exit_codes: tuple[int, ...],
-) -> None:
-    """Assert pipeline wait result failure metadata."""
-    assert result.failure_index == failure_index
-    assert result.exit_codes == exit_codes
-
-
-@dc.dataclass(frozen=True)
-class _FailFastScenario:
-    """Configuration for a fail-fast pipeline test scenario."""
-
-    exit_codes: tuple[int, int, int]
-    ready_stages: frozenset[int]
-    expected_failure_index: int
-    expected_exit_codes: tuple[int, int, int]
-    terminated_stages: frozenset[int]
-
-
-@pytest.mark.parametrize(
-    "scenario",
-    [
-        pytest.param(
-            _FailFastScenario(
-                exit_codes=(7, 0, 0),
-                ready_stages=frozenset([0]),
-                expected_failure_index=0,
-                expected_exit_codes=(7, -15, -15),
-                terminated_stages=frozenset([1, 2]),
-            ),
-            id="early-stage-failure-terminates-downstream",
-        ),
-        pytest.param(
-            _FailFastScenario(
-                exit_codes=(0, 3, 0),
-                ready_stages=frozenset([1]),
-                expected_failure_index=1,
-                expected_exit_codes=(-15, 3, -15),
-                terminated_stages=frozenset([0, 2]),
-            ),
-            id="middle-stage-failure-terminates-downstream",
-        ),
-        pytest.param(
-            _FailFastScenario(
-                exit_codes=(0, 0, 5),
-                ready_stages=frozenset([0, 1, 2]),
-                expected_failure_index=2,
-                expected_exit_codes=(0, 0, 5),
-                terminated_stages=frozenset(),
-            ),
-            id="last-stage-failure-no-termination",
-        ),
-    ],
-)
-def test_wait_for_pipeline_fail_fast_scenarios(
-    scenario: _FailFastScenario,
-) -> None:
-    """Validate fail-fast termination behaviour across different failure scenarios.
-
-    Tests that:
-    - Early stage failures terminate all downstream stages
-    - Middle stage failures terminate all downstream stages
-    - Final stage failures record failure index without terminating others
-    """
-    p0, p1, p2, result = asyncio.run(
-        _exercise_wait_for_pipeline(
-            exit_codes=scenario.exit_codes,
-            ready_stages=scenario.ready_stages,
-        ),
-    )
-
-    _assert_pipeline_failure(
-        result,
-        failure_index=scenario.expected_failure_index,
-        exit_codes=scenario.expected_exit_codes,
-    )
-
-    for idx, process in enumerate([p0, p1, p2]):
-        _assert_stage_terminated(
-            process,
-            should_terminate=(idx in scenario.terminated_stages),
-        )
-
-
-@pytest.mark.parametrize(
-    ("capture", "expected_stdout", "expected_stderr"),
-    [
-        pytest.param(True, "INTERMEDIATE", "", id="capture-final-stdio"),
-        pytest.param(False, None, None, id="discard-final-stdio"),
-    ],
-)
-def test_pipeline_stdio_policy_streams_intermediate_stdout_end_to_end(
-    *,
-    capture: bool,
-    expected_stdout: str | None,
-    expected_stderr: str | None,
-) -> None:
-    """Pipeline execution streams intermediate stdout and applies final capture."""
-    catalogue, python_program = python_catalogue()
-    python = sh.make(python_program, catalogue=catalogue)
-
-    producer = python("-c", "import sys; sys.stdout.write('intermediate')")
-    transformer = python(
-        "-c",
-        "import sys; sys.stdout.write(sys.stdin.read().upper())",
-    )
-
-    with scoped(ScopeConfig(allowlist=frozenset([python_program]))):
-        result = (producer | transformer).run_sync(
-            output=RunOutputOptions(capture=capture),
-        )
-
-    assert result.stdout == expected_stdout, (
-        f"capture={capture}: result.stdout mismatch"
-    )
-    assert len(result.stages) == 2, f"capture={capture}: result.stages length mismatch"
-    assert result.stages[0].stdout is None, (
-        f"capture={capture}: stage 0 stdout mismatch"
-    )
-    assert result.stages[0].stderr == expected_stderr, (
-        f"capture={capture}: stage 0 stderr mismatch"
-    )
-    assert result.stages[0].exit_code == 0, (
-        f"capture={capture}: stage 0 exit_code mismatch"
-    )
-    assert result.stages[1].stdout == expected_stdout, (
-        f"capture={capture}: stage 1 stdout mismatch"
-    )
-    assert result.stages[1].stderr == expected_stderr, (
-        f"capture={capture}: stage 1 stderr mismatch"
-    )
-    assert result.stages[1].exit_code == 0, (
-        f"capture={capture}: stage 1 exit_code mismatch"
-    )
